backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15
psycopg2-binary>=2.9
django-cors-headers>=4.3
PyJWT>=2.8
python-dotenv>=1.0
Pillow>=10.0
